Fix double-zoom glitch
Previously, calls to zoom() were made before zoomMode was updated when a button was clicked, meaning if you were zooming in (zoomMode == 2) and clicked on the zoom out button (which sets zoomMode to 0.5) you would be zoomed in once more before the zoomMode switched. This is no longer the case.